refactor(frontend): Move ExamplesService to EditorModule
